scrutinizer fix issues introduced by me

+++ lib/class/share.class.php
@@ -212,7 +212,7 @@ class Share extends database_object
     public function show_action_buttons()
     {
         if ($this->id) {
-            if ($GLOBALS['user']->has_access('75') || $this->user == User::get_user_id()) {
+            if ($GLOBALS['user']->has_access('75') || $this->user == (int) User::get_user_id()) {
                 echo "<a id=\"edit_share_ " . $this->id . "\" onclick=\"showEditDialog('share_row', '" . $this->id . "', 'edit_share_" . $this->id . "', '" . T_('Share edit') . "', 'share_')\">" . UI::get_icon('edit', T_('Edit')) . "</a>";
                 echo "<a href=\"" . AmpConfig::get('web_path') . "/share.php?action=show_delete&id=" . $this->id . "\">" . UI::get_icon('delete', T_('Delete')) . "</a>";
             }
