+++ server/user.ajax.php
@@ -36,7 +36,7 @@ switch ($action) {
     case 'flip_follow':
         if (Access::check('interface', 25) && AmpConfig::get('sociable')) {
             $fuser = new User($user_id);
-            if ($fuser->id > 0 && $user_id !== User::get_user_id()) {
+            if ($fuser->id > 0 && $user_id !== (int) User::get_user_id()) {
                 $GLOBALS['user']->toggle_follow($user_id);
                 $results['button_follow_' . $user_id] = $fuser->get_display_follow();
             }
